// hw21/utils/types.ts
export enum PROMO_CATEGORIES {
    SCHOOL = "Школа",
    BIKES = "Велосипеды",
    CONDITIONERS = "Кондиционеры",
    REFRIGERATORS = "Холодильники",
    FREEZERS = "Морозильники",
    SMARTPHONES = "Смартфоны", 
    NOTEBOOKS = "Ноутбуки",
    MATRESSES = "Матрасы",
    DOORS = "Входные двери",
    BOILERS = "Котлы",
    SOFAS = "Диваны"
};

export enum SELECTOR_TYPES {
    XPATH = "xpath", 
    CSS = "css",
    ID = "id",
    LINK_TEXT = "linkText"
};

export enum VIEW_TYPES {
    LIST = "Список",
    GRID = "Сетка"
};

export enum SORT_PARAMETERS {
    POPULAR = "популярности",
    PRICE = "цене",
    SALE = "скидке"
};

export enum PAGES {
    HOME = "HomePage",
    BIKES = "BikesPage"
};